+++ src/jmh/java/org/apache/aurora/benchmark/fakes/FakeDriver.java
@@ -27,12 +27,12 @@ public class FakeDriver extends AbstractIdleService implements Driver {
   }
 
   @Override
-  public void launchTask(Protos.OfferID offerId, Protos.TaskInfo task) {
+  public void launchTask(Protos.OfferID offerId, Protos.TaskInfo task, Protos.Filters filter) {
     // no-op
   }
 
   @Override
-  public void declineOffer(Protos.OfferID offerId) {
+  public void declineOffer(Protos.OfferID offerId, Protos.Filters filters) {
     // no-op
   }